# be/apidocs/swagger.yaml
basePath: /api/
definitions:
  controller.BookAppointmentInput:
    properties:
      dateTime:
        type: string
    type: object
  controller.CreateBarbershopInput:
    properties:
      employees_number:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    required:
    - name
    type: object
  controller.FindBarbershopInput:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      radius:
        type: number
    type: object
...
  controller.ModifyBarberShopInput:
    properties:
      employees:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
...
  controller.SearchInput:
    properties:
      area:
        type: string
    required:
    - area
    type: object
...
  entity.Appointment:
    properties:
      barbershopID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      start:
        type: string
      userID:
        type: string
    type: object
  entity.BarberShop:
    properties:
      address:
        type: string
      appointments:
        items:
          $ref: '#/definitions/entity.Appointment'
        type: array
      employees:
        type: integer
      id:
        type: string
      imageLink:
        type: string
      location:
        $ref: '#/definitions/entity.Location'
      name:
        type: string
      phone:
        type: string
      rating:
        type: number
      reviews:
        items:
          $ref: '#/definitions/entity.Review'
        type: array
    type: object
  entity.GeocodingInfo:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      region:
        type: string
    type: object
  entity.Location:
    properties:
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
  entity.Review:
    properties:
      content:
        type: string
      createdAt:
        type: string
      rating:
        type: integer
      reported:
        type: boolean
    type: object
  entity.Slot:
    properties:
      bookedAppoIntments:
        type: integer
      start:
        type: string
      unavailableEmployees:
        type: integer
    type: object
...
host: localhost:7000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gin Swagger Example API
  version: "1.0"
paths:
  /admin/barbershop:
    post:
      consumes:
      - application/json
      description: Creates a new barbershop with the given name, location and number
        of employees.
      parameters:
      - description: The input parameters for creating a new barbershop
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.CreateBarbershopInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Create a new barbershop
      tags:
      - Barbershop
  /admin/barbershop/{shopid}:
    delete:
      consumes:
      - application/json
      description: Deletes a barbershop by ID.
      parameters:
      - description: The ID of the barbershop to delete
        in: path
        name: shopid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a barbershop by ID
      tags:
      - Barbershop
    put:
      consumes:
      - application/json
      description: Modifies details of a barbershop for the given shop ID.
      parameters:
      - description: The ID of the barbershop
        in: path
        name: shopid
        required: true
        type: string
      - description: The updated barbershop details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.ModifyBarberShopInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Modify details of a barbershop
      tags:
      - Barbershop
...
  /barbershop:
    post:
      consumes:
      - application/json
      description: Finds barbershops within a certain radius of the given coordinates
        and name.
      parameters:
      - description: The input parameters for finding barbershops
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.FindBarbershopInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/entity.BarberShop'
              type: array
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Find barbershops within a certain radius
      tags:
      - Barbershop
  /barbershop/{shopid}:
    get:
      consumes:
      - application/json
      description: Retrieves details of a barbershop for the given shop ID.
      parameters:
      - description: The ID of the barbershop
        in: path
        name: shopid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/entity.BarberShop'
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve details of a barbershop
      tags:
      - Barbershop
  /barbershop/{shopid}/appointment:
    post:
      consumes:
      - application/json
      description: Books a new appointment for the current user.
      parameters:
      - description: The ID of the barbershop
        in: path
        name: shopid
        required: true
        type: string
      - description: The appointment details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.BookAppointmentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Book a new appointment
      tags:
      - Appointment
  /barbershop/{shopid}/appointment/{appointmentid}:
    delete:
      consumes:
      - application/json
      description: Deletes an appointment at a specific barbershop
      parameters:
      - description: ID of the barbershop
        in: path
        name: shopid
        required: true
        type: string
      - description: ID of the appointment
        in: path
        name: appointmentid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Deletes an appointment
      tags:
      - Appointment
  /barbershop/{shopid}/calendar:
    get:
      consumes:
      - application/json
      description: Gets the calendar for a barbershop by ID.
      parameters:
      - description: The ID of the barbershop to get the calendar for
        in: path
        name: shopid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/entity.Slot'
              type: array
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get the calendar for a barbershop by ID
      tags:
      - Barbershop
...
  /geocoding/search:
    post:
      consumes:
      - application/json
      description: Given some information about a place returns a list of possible
        location with some other information about them in order of importance
      parameters:
      - description: Area to search
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.SearchInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/entity.GeocodingInfo'
              type: array
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Given an area returns an array of geocodes
      tags:
      - Geocoding
...
  /user/self/appointment:
    delete:
      consumes:
      - application/json
      description: Deletes the appointment of the current user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Deletes the current user's appointment
      tags:
      - Appointment
schemes:
- http
swagger: "2.0"